py/requirements.txt
google-cloud-bigquery
google-cloud-logging
functions-framework==1.3.0
google-cloud-pubsub
